entity 구조 변경. 코드 리팩토링.

+++ src/main/java/com/odi/parser/model/asset/AssetNode.java
@@ -1,11 +1,8 @@
 package com.odi.parser.model.asset;
 
-import com.odi.parser.model.Relation;
 import com.odi.parser.model.enums.AssetType;
 import com.odi.parser.model.enums.RelationType;
 import lombok.Data;
-
-import java.sql.Date;
 
 @Data
 public class AssetNode {
@@ -13,8 +10,8 @@ public class AssetNode {
     private RelationType relationType;
     private String name;
     private String type;
-    private Integer lastPrice;
-    private Integer price;
+    private Long lastPrice;
+    private Long price;
     private String description;
     private String reason;
 
@@ -26,18 +23,18 @@ public class AssetNode {
     public void setLastPrice(String lastPrice) {
         String value = lastPrice.replace(",","").replace("-", "");
         if(value.isEmpty()) {
-            this.lastPrice = 0;
+            this.lastPrice = Long.valueOf(0);
         } else {
-            this.lastPrice = Integer.valueOf(value);
+            this.lastPrice = Long.valueOf(value);
         }
     }
 
     public void setPrice(String price) {
         String value = price.replace(",","").replace("-", "");
         if(value.isEmpty()) {
-            this.price = 0;
+            this.price = Long.valueOf(0);
         } else {
-            this.price = Integer.valueOf(value);
+            this.price = Long.valueOf(value);
         }
     }
 }
